Move rustc_middle::middle::cstore to rustc_crate.
