+++ compiler/rustc_codegen_cranelift/src/driver/mod.rs
@@ -3,8 +3,8 @@
 
 use std::any::Any;
 
+use rustc_crate::cstore::EncodedMetadata;
 use rustc_middle::middle::codegen_fn_attrs::{Linkage as RLinkage, Visibility};
-use rustc_middle::middle::cstore::EncodedMetadata;
 use rustc_middle::mir::mono::MonoItem;
 
 use crate::prelude::*;
