+++ compiler/rustc_driver/Cargo.toml
@@ -14,6 +14,7 @@ tracing-subscriber = { version = "0.2.13", default-features = false, features = 
 tracing-tree = "0.1.6"
 rustc_middle = { path = "../rustc_middle" }
 rustc_ast_pretty = { path = "../rustc_ast_pretty" }
+rustc_crate = { path = "../rustc_crate" }
 rustc_target = { path = "../rustc_target" }
 rustc_lint = { path = "../rustc_lint" }
 rustc_data_structures = { path = "../rustc_data_structures" }
