+++ compiler/rustc_interface/Cargo.toml
@@ -15,6 +15,7 @@ smallvec = { version = "1.0", features = ["union", "may_dangle"] }
 rustc_ast = { path = "../rustc_ast" }
 rustc_attr = { path = "../rustc_attr" }
 rustc_builtin_macros = { path = "../rustc_builtin_macros" }
+rustc_crate = { path = "../rustc_crate" }
 rustc_expand = { path = "../rustc_expand" }
 rustc_parse = { path = "../rustc_parse" }
 rustc_session = { path = "../rustc_session" }
