+++ compiler/rustc_plugin_impl/src/load.rs
@@ -2,9 +2,9 @@
 
 use crate::Registry;
 use rustc_ast::Crate;
+use rustc_crate::cstore::MetadataLoader;
 use rustc_errors::struct_span_err;
 use rustc_metadata::locator;
-use rustc_middle::middle::cstore::MetadataLoader;
 use rustc_session::Session;
 use rustc_span::symbol::{sym, Ident};
 use rustc_span::Span;
