+++ src/librustdoc/lib.rs
@@ -31,6 +31,7 @@ extern crate tracing;
 extern crate rustc_ast;
 extern crate rustc_ast_pretty;
 extern crate rustc_attr;
+extern crate rustc_crate;
 extern crate rustc_data_structures;
 extern crate rustc_driver;
 extern crate rustc_errors;
